re-struct the directories
