--- include/gc-dialects/Linalgx/LinalgxOps.h
@@ -1,17 +1,0 @@
-//===- LinalgxOps.h - linalgx dialect ops -----------------------*- C++ -*-===//
-//
-// This file is licensed under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-//===----------------------------------------------------------------------===//
-
-#ifndef GC_DIALECTS_LINALGXOPS_H
-#define GC_DIALECTS_LINALGXOPS_H
-
-#include "mlir/IR/OpDefinition.h"
-
-#define GET_OP_CLASSES
-#include "gc-dialects/Linalgx/LinalgxOps.h.inc"
-
-#endif // GC_DIALECTS_LINALGXOPS_H
