--- include/gc-dialects/Microkernel/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_mlir_dialect(MicrokernelOps microkernel)
-add_mlir_doc(MicrokernelOps MicrokernelOps gc-dialects/Microkernel/ -gen-op-doc)
-add_mlir_doc(MicrokernelDialect MicrokernelDialect gc-dialects/Microkernel/ -gen-dialect-doc)
