+++ include/gc/Dialects/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_subdirectory(OnednnGraph)
+add_subdirectory(Microkernel)
+add_subdirectory(Linalgx)
