+++ include/gc/Dialects/Linalgx/LinalgxDialect.h
@@ -1,0 +1,16 @@
+//===- LinalgxDialect.h - linalgx dialect -----------------------*- C++ -*-===//
+//
+// This file is licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef GC_DIALECTS_LINALGXDIALECT_H
+#define GC_DIALECTS_LINALGXDIALECT_H
+
+#include "mlir/IR/Dialect.h"
+
+#include "gc/Dialects/Linalgx/LinalgxOpsDialect.h.inc"
+
+#endif // GC_DIALECTS_LINALGXDIALECT_H
