+++ include/gc/Dialects/Microkernel/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_mlir_dialect(MicrokernelOps microkernel)
+add_mlir_doc(MicrokernelOps MicrokernelOps gc/Dialects/Microkernel/ -gen-op-doc)
+add_mlir_doc(MicrokernelDialect MicrokernelDialect gc/Dialects/Microkernel/ -gen-dialect-doc)
