+++ include/gc/Dialects/Microkernel/MicrokernelDialect.h
@@ -1,0 +1,16 @@
+//===- MicrokernelDialect.h - microkernel dialect ---------------*- C++ -*-===//
+//
+// This file is licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef GC_DIALECTS_MICROKERNELDIALECT_H
+#define GC_DIALECTS_MICROKERNELDIALECT_H
+
+#include "mlir/IR/Dialect.h"
+
+#include "gc/Dialects/Microkernel/MicrokernelOpsDialect.h.inc"
+
+#endif // GC_DIALECTS_MICROKERNELDIALECT_H
