+++ include/gc/Dialects/OnednnGraph/OnednnGraphDialect.h
@@ -1,0 +1,16 @@
+//===- OnednnGraphDialect.h - OneDNN input dialect --------------*- C++ -*-===//
+//
+// This file is licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef GC_DIALECTS_ONEDNNGRAPHDIALECT_H
+#define GC_DIALECTS_ONEDNNGRAPHDIALECT_H
+
+#include "mlir/IR/Dialect.h"
+
+#include "gc/Dialects/OnednnGraph/OnednnGraphOpsDialect.h.inc"
+
+#endif // GC_DIALECTS_ONEDNNGRAPHDIALECT_H
