--- lib/gc-dialects/CMakeLists.txt
@@ -1,5 +1,0 @@
-add_subdirectory(Linalgx)
-add_subdirectory(Microkernel)
-add_subdirectory(OnednnGraph)
-
-add_subdirectory(Transforms)
