+++ lib/gc/Dialects/Linalgx/CMakeLists.txt
@@ -1,0 +1,13 @@
+add_mlir_dialect_library(MLIRLinalgx
+    LinalgxDialect.cpp
+    LinalgxOps.cpp
+    
+    ADDITIONAL_HEADER_DIRS
+    ${PROJECT_SOURCE_DIR}/include/gc/Dialects/Linalgx
+    
+    DEPENDS
+    MLIRLinalgxOpsIncGen
+    
+    LINK_LIBS PUBLIC
+    MLIRIR
+)
