+++ lib/gc/Dialects/Linalgx/LinalgxDialect.cpp
@@ -1,0 +1,20 @@
+//===- LinalgxDialect.h - linalgx dialect -----------------------*- C++ -*-===//
+//
+// This file is licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#include "gc/Dialects/Linalgx/LinalgxDialect.h"
+#include "gc/Dialects/Linalgx/LinalgxOps.h"
+
+using namespace mlir;
+using namespace mlir::linalgx;
+
+void LinalgxDialect::initialize() {
+  addOperations<
+#define GET_OP_LIST
+#include "gc/Dialects/Linalgx/LinalgxOps.cpp.inc"
+      >();
+}
