+++ lib/gc/Dialects/Microkernel/CMakeLists.txt
@@ -1,0 +1,13 @@
+add_mlir_dialect_library(MLIRMicrokernel
+    MicrokernelDialect.cpp
+    MicrokernelOps.cpp
+    
+    ADDITIONAL_HEADER_DIRS
+    ${PROJECT_SOURCE_DIR}/include/gc/Dialects/Microkernel
+    
+    DEPENDS
+    MLIRMicrokernelOpsIncGen
+    
+    LINK_LIBS PUBLIC
+    MLIRIR
+)
