+++ lib/gc/Dialects/Microkernel/MicrokernelDialect.cpp
@@ -1,0 +1,20 @@
+//===- MicrokernelDialect.h - microkernel dialect ---------------*- C++ -*-===//
+//
+// This file is licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#include "gc/Dialects/Microkernel/MicrokernelDialect.h"
+#include "gc/Dialects/Microkernel/MicrokernelOps.h"
+
+using namespace mlir;
+using namespace mlir::microkernel;
+
+void MicrokernelDialect::initialize() {
+  addOperations<
+#define GET_OP_LIST
+#include "gc/Dialects/Microkernel/MicrokernelOps.cpp.inc"
+      >();
+}
