+++ lib/gc/Dialects/OnednnGraph/CMakeLists.txt
@@ -1,0 +1,13 @@
+add_mlir_dialect_library(MLIROnednnGraph
+    OnednnGraphDialect.cpp
+    OnednnGraphOps.cpp
+    
+    ADDITIONAL_HEADER_DIRS
+    ${PROJECT_SOURCE_DIR}/include/gc/Dialects/OnednnGraph
+    
+    DEPENDS
+    MLIROnednnGraphOpsIncGen
+    
+    LINK_LIBS PUBLIC
+    MLIRIR
+)
